KYCFirst/requirements.txt
asgiref==3.6.0
Babel==2.11.0
Django==4.1.6
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
phonenumbers==8.13.6
Pillow==9.4.0
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
